refactor(app, shared-data, components): move ProtocolDeck and utils
