+++ app/src/organisms/Desktop/ProtocolsLanding/ProtocolCard.tsx
@@ -17,6 +17,7 @@ import {
   ModuleIcon,
   OVERFLOW_WRAP_ANYWHERE,
   POSITION_ABSOLUTE,
+  ProtocolDeck,
   SIZE_2,
   SIZE_3,
   SPACING,
@@ -33,7 +34,6 @@ import {
 } from '@opentrons/shared-data'
 
 import { InstrumentContainer } from '/app/atoms/InstrumentContainer'
-import { ProtocolDeck } from '/app/organisms/ProtocolDeck'
 import { getIsProtocolAnalysisInProgress } from '/app/redux/protocol-storage'
 import { getAnalysisStatus } from '/app/transformations/analysis'
 import { getProtocolUsesGripper } from '/app/transformations/commands'
