+++ app/src/organisms/InterventionModal/StackerEmptyInterventionContent.tsx
@@ -10,10 +10,12 @@ import {
   SPACING,
 } from '@opentrons/components'
 import { useRunCurrentState } from '@opentrons/react-api-client'
-import { getLoadedLabwareDefinitionsByUri } from '@opentrons/shared-data'
+import {
+  getLoadedLabwareDefinitionsByUri,
+  getStackerLocationFromSlotName,
+} from '@opentrons/shared-data'
 
 import { InterventionInfo } from '/app/molecules/InterventionModal/InterventionContent'
-import { getStackerLocationFromSlotName } from '/app/transformations/commands'
 
 import { InterventionCommandMessage } from './InterventionCommandMessage'
