+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupLabware/SetupLabwareStackView.tsx
@@ -19,23 +19,23 @@ import {
   truncateString,
 } from '@opentrons/components'
 import {
+  getLabwareDefinitionsByURIForProtocol,
   getSchema2CornerOffsetFromSlot,
   getSchema2Dimensions,
+  getWellFillFromLabwareId,
 } from '@opentrons/shared-data'
 
 import { LabwareStackContents } from '/app/molecules/LabwareStackContents'
 import { ODDBackButton } from '/app/molecules/ODDBackButton'
-import { getWellFillFromLabwareId } from '/app/organisms/ProtocolDeck'
-import { getLabwareDefinitionsByURIForProtocol } from '/app/transformations/commands'
 
 import { LabwareLiquidsDetailModal } from './LabwareLiquidsDetailModal'
 
-import type { CompletedProtocolAnalysis } from '@opentrons/shared-data'
 import type {
+  CompletedProtocolAnalysis,
   LabwareByLiquidId,
   LabwareInStack,
   StackItem,
-} from '/app/transformations/commands'
+} from '@opentrons/shared-data'
 
 const LabwareThumbnail = styled.svg`
   transform: scale(1, -1);
