--- app/src/organisms/ProtocolDeck/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './getStandardDeckViewLayerBlockList'
-export * from './getWellFillFromLabwareId'
