+++ app/src/pages/ODD/QuickTransferDetails/Deck.tsx
@@ -1,12 +1,10 @@
 import last from 'lodash/last'
 
-import { Flex, SPACING } from '@opentrons/components'
+import { Flex, ProtocolDeck, SPACING } from '@opentrons/components'
 import {
   useProtocolAnalysisAsDocumentQuery,
   useProtocolQuery,
 } from '@opentrons/react-api-client'
-
-import { ProtocolDeck } from '/app/organisms/ProtocolDeck'
 
 export const Deck = (props: { transferId: string }): JSX.Element => {
   const { data: transferData } = useProtocolQuery(props.transferId)
