+++ app/src/transformations/commands/transformations/__tests__/getLabwareInforByLiquidId.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { getLabwareInfoByLiquidId } from '../getLabwareInfoByLiquidId'
+import { getLabwareInfoByLiquidId } from '@opentrons/shared-data/js/helpers/getLabwareInfoByLiquidId'
 
 import type { RunTimeCommand } from '@opentrons/shared-data'
