+++ app/src/transformations/commands/transformations/index.ts
@@ -9,7 +9,4 @@ export * from './getSlotLabwareDefinition'
 export * from './getTipracksVisited'
 export * from './getPickUpTipCommandsWithPipette'
 export * from './getRequiredLabwareDetailsFromLoadCommands'
-export * from './getStackedItemsOnStartingDeck'
-export * from './getLabwareDefinitionsByURIForProtocol'
-export * from './getLabwareInfoByLiquidId'
 export * from './getInitiallyLoadedLabwareByAdapter'
