+++ components/src/organisms/index.ts
@@ -1,5 +1,6 @@
-export * from './DeckLabelSet'
-export * from './Toolbox'
-export * from './ProtocolTimelineScrubber'
 export * from './CommandText'
+export * from './DeckLabelSet'
 export * from './FixtureOption'
+export * from './ProtocolDeck'
+export * from './ProtocolTimelineScrubber'
+export * from './Toolbox'
